[FLINK-15768] [Client/Job Submission] Consolidate executor related class into same module in flink-clients
