+++ flink-clients/src/main/java/org/apache/flink/client/deployment/executors/RemoteExecutor.java
@@ -19,7 +19,6 @@
 package org.apache.flink.client.deployment.executors;
 
 import org.apache.flink.annotation.Internal;
-import org.apache.flink.client.deployment.AbstractSessionClusterExecutor;
 import org.apache.flink.client.deployment.StandaloneClientFactory;
 import org.apache.flink.client.deployment.StandaloneClusterId;
 import org.apache.flink.core.execution.PipelineExecutor;
